Resolved Conflicts:
	daemon/src/account.h
	daemon/src/audio/codecs/celtcodec.cpp
	daemon/src/audio/pulseaudio/pulselayer.cpp
	daemon/src/config/sfl_config.cpp
	daemon/src/config/yamlemitter.cpp
	daemon/src/config/yamlemitter.h
	daemon/src/config/yamlparser.cpp
	daemon/src/main.cpp
	daemon/src/preferences.cpp
	daemon/src/preferences.h
	daemon/src/sip/sdp.cpp
	daemon/src/sip/sdp.h
	daemon/src/sip/sipaccount.cpp
	daemon/src/sip/sipvoiplink.cpp
	daemon/test/audiolayertest.h
	daemon/test/gaincontroltest.cpp
	daemon/test/sdptest.cpp
	gnome/configure.ac
	gnome/src/dbus/dbus.c

+++ daemon/src/audio/audiortp/audio_srtp_session.h
@@ -153,22 +153,22 @@ class AudioSrtpSession : public AudioSymmetricRtpSession {
         uint8 localMasterKey_[16];
 
         /** local master key length in byte */
-        int localMasterKeyLength_;
+        size_t localMasterKeyLength_;
 
         uint8 localMasterSalt_[14];
 
         /** local master salt length in byte */
-        int localMasterSaltLength_;
+        size_t localMasterSaltLength_;
 
         uint8 remoteMasterKey_[16];
 
         /** remote master key length in byte */
-        int remoteMasterKeyLength_;
+        size_t remoteMasterKeyLength_;
 
         uint8 remoteMasterSalt_[14];
 
         /** remote master salt length in byte */
-        int remoteMasterSaltLength_;
+        size_t remoteMasterSaltLength_;
 
         /** Used to make sure remote crypto context not initialized wice. */
         bool remoteOfferIsSet_;
